Add conor foundry config, use new acme defaults

+++ common/sysconfig.nix
@@ -18,7 +18,9 @@
     findutils = pkgs.findutils;
   in ''
     umask 0027
-    ${findutils}/bin/find ~/.vscode-server -type f -name node \( -execdir rm '{}' \; -and -execdir ln -s '${node}/bin/node' '{}' \; \)
+    if test -e ~/.vscode-server; then
+      ${findutils}/bin/find ~/.vscode-server -type f -name node \( -execdir rm '{}' \; -and -execdir ln -s '${node}/bin/node' '{}' \; \)
+    fi
   '';
 
   # Enable rsyslog
